+++ bika/lims/decorators.py
@@ -23,6 +23,7 @@ import json
 import os
 import threading
 import time
+import transaction
 from functools import wraps
 
 from bika.lims import api
@@ -157,3 +158,17 @@ def synchronized(max_connections=2, verbose=0):
 
         return wrapper
     return inner
+
+
+def readonly_transaction(func):
+    """Decorator to doom the current transaction
+    https://transaction.readthedocs.io/en/latest/doom.html#dooming-transactions
+    """
+    @wraps(func)
+    def decorator(self, *args, **kwargs):
+        logger.info("*** READONLY TRANSACTION: '{}.{}' ***".
+                    format(self.__class__.__name__, self.__name__))
+        tx = transaction.get()
+        tx.doom()
+        return func(self, *args, **kwargs)
+    return decorator